# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/log/": {
            "get": {
                "description": "Возвращает логи приложения",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Получить лог",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Формат вывода (str - по умолчанию, json, raw)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/log/clear/": {
            "post": {
                "description": "Очищает файл логов приложения",
                "produces": [
                    "plain/text"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Очистить лог",
                "responses": {
                    "200": {
                        "description": "Логи очищены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при очистке логов",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Очищает файл логов приложения",
                "produces": [
                    "plain/text"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Очистить лог",
                "responses": {
                    "200": {
                        "description": "Логи очищены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при очистке логов",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/get/tag/": {
            "get": {
                "description": "Возвращает значение тега на выбранную дату.",
                "produces": [
                    "plain/text json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Получить значение тега",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Наименование тега",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата начала периода",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания периода",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Функция группировки (avg, sum, count, min, max)",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Количество значений",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Округление, знаков после запятой (по умолчанию 2)",
                        "name": "round",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Формат вывода (str - по умолчанию, json, raw)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/get/tag/down/": {
            "get": {
                "description": "Возвращает дату и время отключения оборудования",
                "produces": [
                    "plain/text"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Получить даты отключения оборудования",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Наименование тега",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата начала периода",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания периода",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Номер отключения после даты начала (0 - первое отключение)",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/get/tag/list/": {
            "get": {
                "description": "Возвращает список всех тегов по маске",
                "produces": [
                    "plain/text"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Получить список тегов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Маска поиска",
                        "name": "like",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Формат вывода (str - по умолчанию, json, raw)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/get/tag/up/": {
            "get": {
                "description": "Возвращает дату и время включения оборудования",
                "produces": [
                    "plain/text"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Получить даты включения оборудования",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Наименование тега",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата начала периода",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата окончания периода",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Номер включения после даты начала (0 - первое включение)",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tag/decode/": {
            "get": {
                "description": "Возвращает расшифровку имени тега",
                "produces": [
                    "plain/text json"
                ],
                "tags": [
                    "Tag"
                ],
                "summary": "Получить расшифровку имени тега",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Наименование тега",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Формат вывода (str - по умолчанию, json, raw)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    }
}